@@ -8,7 +8,7 @@ dynamic = ["version"]
 description = "Safe Ecosystem Foundation utilities for Ethereum projects"
 readme = "README.rst"
 license = "MIT"
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 authors = [
     { name = "Uxío", email = "uxio@safe.global" },
 ]
@@ -27,14 +27,15 @@ classifiers = [
     "Framework :: Django :: 2.0",
     "Framework :: Django :: 3.0",
     "Framework :: Django :: 4.0",
+    "Framework :: Django :: 5.0",
     "Intended Audience :: Developers",
     "License :: OSI Approved :: MIT License",
     "Operating System :: OS Independent",
     "Programming Language :: Python",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Programming Language :: Python :: Implementation :: CPython",
     "Topic :: Internet :: WWW/HTTP",
     "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
@@ -44,7 +45,7 @@ dependencies = [
     "py-evm>=0.7.0a1",
     "requests>=2",
     "safe-pysha3>=1.0.0",
-    "web3>=6.2.0",
+    "web3>=6.2.0,<7.0.0",
 ]
 
 [project.optional-dependencies]
@@ -122,7 +123,7 @@ sections = [
 ]
 
 [tool.mypy]
-python_version = "3.12"
+python_version = "3.13"
 check_untyped_defs = true
 ignore_missing_imports = true
 warn_unused_ignores = true
