@@ -1,1 +1,1 @@
-VERSION = "6.0.0b37"
+VERSION = "6.4.0"
