@@ -142,6 +142,7 @@ class EthereumNetwork(Enum):
     FACTORY_127_MAINNET = 127
     HUOBI_ECO_CHAIN_MAINNET = 128
     INNOVATOR_CHAIN = 129
+    UNICHAIN = 130
     ENGRAM_TESTNET = 131
     NAMEFI_CHAIN_MAINNET = 132
     HASHKEY_CHAIN_TESTNET = 133
@@ -154,8 +155,10 @@ class EthereumNetwork(Enum):
     ETERNAL_MAINNET = 140
     OPENPIECE_TESTNET = 141
     DAX_CHAIN = 142
+    MONAD_MAINNET = 143
     PHI_NETWORK_V2 = 144
     SORAAI_TESTNET = 145
+    SONIC_MAINNET = 146
     FLAG_MAINNET = 147
     SHIMMEREVM = 148
     SIX_PROTOCOL_TESTNET = 150
@@ -165,26 +168,31 @@ class EthereumNetwork(Enum):
     REDBELLY_NETWORK_TGE = 154
     TENET_TESTNET = 155
     OEBLOCK_TESTNET = 156
-    PUPPYNET_SHIBARIUM = 157
+    PUPPYNET = 157
     ROBURNA_MAINNET = 158
     ROBURNA_TESTNET = 159
     ARMONIA_EVA_CHAIN_MAINNET = 160
     ARMONIA_EVA_CHAIN_TESTNET = 161
     LIGHTSTREAMS_TESTNET = 162
     LIGHTSTREAMS_MAINNET = 163
-    OMNI_OMEGA = 164
+    OMNI_OMEGA_TESTNET = 164
     OMNI_TESTNET_DEPRECATED = 165
     OMNI = 166
     ATOSHI_TESTNET = 167
     AIOZ_NETWORK = 168
     MANTA_PACIFIC_MAINNET = 169
     HOO_SMART_CHAIN_TESTNET = 170
+    CO2E_LEDGER = 171
     LATAM_BLOCKCHAIN_RESIL_TESTNET = 172
+    OTC = 175
     DC_MAINNET = 176
+    HASHKEY_CHAIN = 177
     ABEY_TESTNET = 178
     ABEY_MAINNET = 179
     AME_CHAIN_MAINNET = 180
     WATERFALL_NETWORK = 181
+    IOST_MAINNET = 182
+    ETHERNITY = 183
     DOJIMA_TESTNET = 184
     MINT_MAINNET = 185
     SEELE_MAINNET = 186
@@ -203,6 +211,7 @@ class EthereumNetwork(Enum):
     MOAC_TESTNET = 201
     EDGELESS_TESTNET = 202
     OPBNB_MAINNET = 204
+    EKAASH = 205
     VINUCHAIN_TESTNET = 206
     VINUCHAIN_NETWORK = 207
     STRUCTX_MAINNET = 208
@@ -220,21 +229,27 @@ class EthereumNetwork(Enum):
     VIRIDIS_TESTNET = 224
     LACHAIN_MAINNET = 225
     LACHAIN_TESTNET = 226
+    PROM = 227
     MIND_NETWORK_MAINNET = 228
     SWAPDEX = 230
+    ETHERNITY_TESTNET = 233
     PROTOJUMBO_TESTNET = 234
     DEAMCHAIN_TESTNET = 236
     BLAST_MAINNET = 238
+    CRONOS_ZKEVM_TESTNET = 240
     PLINGA_MAINNET = 242
     ENERGY_WEB_CHAIN = 246
     OASYS_MAINNET = 248
     FANTOM_OPERA = 250
+    GLIDE_L1_PROTOCOL_XP = 251
     FRAXTAL = 252
+    GLIDE_L2_PROTOCOL_XP = 253
     SWAN_CHAIN_MAINNET = 254
     KROMA = 255
     HUOBI_ECO_CHAIN_TESTNET = 256
     SETHEUM = 258
     NEONLINK_MAINNET = 259
+    GURU_NETWORK = 260
     GURU_NETWORK_TESTNET = 261
     SUR_BLOCKCHAIN_NETWORK = 262
     NEURA = 266
@@ -244,11 +259,12 @@ class EthereumNetwork(Enum):
     EGONCOIN_MAINNET = 271
     LACHAIN = 274
     XFAIR_AI_MAINNET = 278
-    BPX_BLOCKCHAIN = 279
+    BPX_CHAIN = 279
     ZKSYNC_ERA_GOERLI_TESTNET_DEPRECATED = 280
-    CRONOS_ZKEVM_TESTNET = 282
+    DEPRECATED_CRONOS_ZKEVM_TESTNET = 282
     BOBA_NETWORK = 288
     ORDERLY_MAINNET = 291
+    DAVINCI = 293
     HEDERA_MAINNET = 295
     HEDERA_TESTNET = 296
     HEDERA_PREVIEWNET = 297
@@ -268,8 +284,9 @@ class EthereumNetwork(Enum):
     KCC_TESTNET = 322
     COSVM_MAINNET = 323
     ZKSYNC_MAINNET = 324
-    GRVT_MAINNET = 325
-    GRVT_SEPOLIA_TESTNET = 326
+    GRVT_EXCHANGE = 325
+    GRVT_EXCHANGE_TESTNET = 326
+    OMAX_TESTNET = 332
     WEB3Q_MAINNET = 333
     DFK_CHAIN_TEST = 335
     SHIDEN = 336
@@ -287,8 +304,8 @@ class EthereumNetwork(Enum):
     LISINSKI = 385
     CRONOS_ZKEVM_MAINNET = 388
     CAMDL_TESTNET = 395
-    NEAR_MAINNET = 397
-    NEAR_TESTNET = 398
+    NEAR_PROTOCOL = 397
+    NEAR_PROTOCOL_TESTNET = 398
     NATIV3_MAINNET = 399
     HYPERONCHAIN_TESTNET = 400
     OZONE_CHAIN_TESTNET = 401
@@ -299,6 +316,8 @@ class EthereumNetwork(Enum):
     OPTIMISM_GOERLI_TESTNET = 420
     VIRIDIS_MAINNET = 422
     PGN_PUBLIC_GOODS_NETWORK = 424
+    STENIX_MAINNET = 425
+    THE_WIDOWS_MITE = 426
     ZEETH_CHAIN = 427
     GESO_VERSE = 428
     BOYAA_MAINNET = 434
@@ -307,6 +326,8 @@ class EthereumNetwork(Enum):
     ARZIO_CHAIN = 456
     AREON_NETWORK_TESTNET = 462
     AREON_NETWORK_MAINNET = 463
+    APPCHAIN = 466
+    FORM_NETWORK = 478
     WORLD_CHAIN = 480
     RUPAYA = 499
     CAMINO_C_CHAIN = 500
@@ -321,9 +342,10 @@ class EthereumNetwork(Enum):
     CANDLE = 534
     OPTRUST_MAINNET = 537
     PAWCHAIN_TESTNET = 542
-    TESTNET = 545
+    EVM_ON_FLOW_TESTNET = 545
     VELA1_CHAIN_MAINNET = 555
     TAO_NETWORK = 558
+    PROMETHEUZ_TESTNET = 565
     DOGECHAIN_TESTNET = 568
     ROLLUX_MAINNET = 570
     METACHAIN_MAINNET = 571
@@ -353,12 +375,17 @@ class EthereumNetwork(Enum):
     JUNCACHAIN = 668
     JUNCACHAIN_TESTNET = 669
     KARURA_NETWORK = 686
+    NERO_TESTNET = 689
     REDSTONE = 690
+    MATCHAIN = 698
+    MATCHAIN_TESTNET = 699
     STAR_SOCIAL_TESTNET = 700
     DARWINIA_KOI_TESTNET = 701
     BLOCKCHAIN_STATION_MAINNET = 707
     BLOCKCHAIN_STATION_TESTNET = 708
     HIGHBURY = 710
+    TUCANA = 711
+    BIRDEE_2 = 712
     VRCSCAN_MAINNET = 713
     SHIBARIUM_BETA = 719
     LYCAN_CHAIN = 721
@@ -367,6 +394,7 @@ class EthereumNetwork(Enum):
     CANTO_TESTNET = 740
     VENTION_SMART_CHAIN_TESTNET = 741
     SCRIPT_TESTNET = 742
+    EVM_ON_FLOW = 747
     QL1 = 766
     OPENCHAIN_TESTNET = 776
     CHEAPETH = 777
@@ -385,6 +413,8 @@ class EthereumNetwork(Enum):
     CALLISTO_MAINNET = 820
     CALLISTO_TESTNET_DEPRECATED = 821
     RUNIC_CHAIN_TESTNET = 822
+    DAILY_NETWORK_MAINNET = 824
+    DAILY_NETWORK_TESTNET = 825
     CHECKDOT_BLOCKCHAIN_DEVNET = 831
     TARAXA_MAINNET = 841
     TARAXA_TESTNET = 842
@@ -405,12 +435,15 @@ class EthereumNetwork(Enum):
     DECENTRABONE_LAYER1_TESTNET = 910
     TAPROOT_MAINNET = 911
     RINIA_TESTNET = 917
+    SLERFCHAIN_MAINNET = 918
     MODE_TESTNET = 919
     YIDARK_CHAIN_MAINNET = 927
+    HAUST_MAINNET = 938
     PULSECHAIN_TESTNET = 940
     PULSECHAIN_TESTNET_V2B = 941
     PULSECHAIN_TESTNET_V3 = 942
     PULSECHAIN_TESTNET_V4 = 943
+    SUBTENSOR_EVM_TESTNET = 945
     MUNODE_TESTNET = 956
     LYRA_CHAIN = 957
     BTC20_SMART_CHAIN = 963
@@ -418,6 +451,7 @@ class EthereumNetwork(Enum):
     OORT_MAINNET = 970
     OORT_HUYGENS = 971
     OORT_ASCRAEUS = 972
+    PALM_SMART_CHAIN = 973
     NEPAL_BLOCKCHAIN_NETWORK = 977
     ETHXY_TESTNET = 979
     TOP_MAINNET_EVM = 980
@@ -426,13 +460,16 @@ class EthereumNetwork(Enum):
     TOP_MAINNET = 989
     ELIBERTY_MAINNET = 990
     _5IRECHAIN_MAINNET = 995
-    _5IRECHAIN_THUNDER = 997
+    BIFROST_POLKADOT_MAINNET = 996
+    _5IRECHAIN_THUNDER_TESTNET = 997
     LUCKY_NETWORK = 998
     WANCHAIN_TESTNET = 999
     GTON_MAINNET = 1000
-    KAIA_TESTNET_KAIROS = 1001
+    KAIA_KAIROS_TESTNET = 1001
     TECTUM_EMISSION_TOKEN = 1003
     T_EKTA = 1004
+    LEMONCHAINTESTNET = 1005
+    LEMONCHAIN = 1006
     NEWTON_TESTNET = 1007
     EURUS_MAINNET = 1008
     JUMBOCHAIN_MAINNET = 1009
@@ -443,6 +480,7 @@ class EthereumNetwork(Enum):
     CLOVER_TESTNET = 1023
     CLV_PARACHAIN = 1024
     BITTORRENT_CHAIN_TESTNET = 1028
+    BITTORRENT_CHAIN_DONAU = 1029
     CONFLUX_ESPACE = 1030
     PROXY_NETWORK_TESTNET = 1031
     BRONOS_TESTNET = 1038
@@ -463,6 +501,7 @@ class EthereumNetwork(Enum):
     WEMIX3_0_MAINNET = 1111
     WEMIX3_0_TESTNET = 1112
     B2_HUB_TESTNET = 1113
+    CORE_BLOCKCHAIN_TESTNET2 = 1114
     CORE_BLOCKCHAIN_TESTNET = 1115
     CORE_BLOCKCHAIN_MAINNET = 1116
     DOGCOIN_MAINNET = 1117
@@ -477,6 +516,7 @@ class EthereumNetwork(Enum):
     FLAG_TESTNET = 1147
     SYMPLEXIA_SMART_CHAIN = 1149
     ORIGIN_TESTNET = 1170
+    LITHEUM_TEST_NETWORK = 1174
     SMART_HOST_TEKNOLOJI_TESTNET = 1177
     CLUBMOS_MAINNET = 1188
     IORA_CHAIN = 1197
@@ -485,12 +525,16 @@ class EthereumNetwork(Enum):
     WORLD_TRADE_TECHNICAL_CHAIN_MAINNET = 1202
     SAITABLOCKCHAIN_SBC = 1209
     CUCKOO_SEPOLIA = 1210
+    ADF_CHAIN_TESTNET = 1212
     POPCATEUM_MAINNET = 1213
     ENTERCHAIN_MAINNET = 1214
+    ADF_CHAIN = 1215
     CYCLE_NETWORK_TESTNET = 1221
     CYCLE_NETWORK_TESTNET_JELLYFISH = 1223
     HYBRID_TESTNET_DEPRECATED = 1224
     HYBRID_TESTNET = 1225
+    BITCOIN_PROTOCOL_TESTNET = 1227
+    CYCLE_NETWORK_TESTNET_CUTTLEFISH = 1228
     EXZO_NETWORK_MAINNET = 1229
     ULTRON_TESTNET = 1230
     ULTRON_MAINNET = 1231
@@ -501,6 +545,7 @@ class EthereumNetwork(Enum):
     OM_PLATFORM_MAINNET = 1246
     DOGETHER_MAINNET = 1248
     CIC_CHAIN_TESTNET = 1252
+    METACCES_TESTNET = 1260
     HALO_MAINNET = 1280
     MOONBEAM = 1284
     MOONRIVER = 1285
@@ -510,6 +555,9 @@ class EthereumNetwork(Enum):
     SWISSTRONIK_TESTNET = 1291
     BOBABEAM = 1294
     BOBABASE_TESTNET = 1297
+    ARGOCHAIN_TESTNET = 1298
+    ARGOCHAIN = 1299
+    UNICHAIN_SEPOLIA_TESTNET = 1301
     DOS_FUJI_SUBNET = 1311
     JAIHO_CHAIN = 1313
     ALYX_MAINNET = 1314
@@ -529,7 +577,7 @@ class EthereumNetwork(Enum):
     AMSTAR_MAINNET = 1388
     JOSEON_MAINNET = 1392
     POLYGON_ZKEVM_TESTNET_OLD = 1402
-    SILICON_ZKEVM_SEPOLIA_TESTNET = 1414
+    SILICON_ZKEVM_SEPOLIA_TESTNET_DEPRECATED = 1414
     POLYGON_ZKEVM_TESTNET_PRE_AUDIT_UPGRADED = 1422
     RIKEZA_NETWORK_MAINNET = 1433
     LIVING_ASSETS_MAINNET = 1440
@@ -538,15 +586,20 @@ class EthereumNetwork(Enum):
     METACHAIN_ISTANBUL = 1453
     CTEX_SCAN_BLOCKCHAIN = 1455
     ZKBASE_MAINNET = 1456
+    VANA = 1480
     VITRUVEO_MAINNET = 1490
     IDOS_GAMES_CHAIN_TESTNET = 1499
     BEVM_CANARY = 1501
     SHERPAX_MAINNET = 1506
     SHERPAX_TESTNET = 1507
+    STORY_TESTNET = 1513
     BEAGLE_MESSAGING_CHAIN = 1515
+    STORY_ODYSSEY_TESTNET = 1516
+    DATACORE_SMART_CHAIN = 1555
     TENET = 1559
     STARCHAIN_TESTNET = 1570
     STARCHAIN = 1578
+    BETHERANCE = 1605
     ETHEREUM_INSCRIPTION_MAINNET = 1617
     CATECOIN_CHAIN_MAINNET = 1618
     ATHEIOS = 1620
@@ -558,6 +611,7 @@ class EthereumNetwork(Enum):
     MINT_TESTNET = 1686
     MINT_SEPOLIA_TESTNET = 1687
     LUDAN_MAINNET = 1688
+    NERO_MAINNET = 1689
     ANYTYPE_EVM_CHAIN = 1701
     TBSI_MAINNET = 1707
     TBSI_TESTNET = 1708
@@ -575,7 +629,11 @@ class EthereumNetwork(Enum):
     CUBE_CHAIN_MAINNET = 1818
     CUBE_CHAIN_TESTNET = 1819
     RUBY_SMART_CHAIN_MAINNET = 1821
+    PLAYBLOCK = 1829
+    VERIFY_TESTNET = 1833
+    HIGHOCTANE_SUBNET = 1853
     TESLAFUNDS = 1856
+    SONEIUM = 1868
     WHITECHAIN = 1875
     GITSHOCK_CARTENZ_TESTNET = 1881
     LIGHTLINK_PHOENIX_MAINNET = 1890
@@ -589,7 +647,12 @@ class EthereumNetwork(Enum):
     SCALIND = 1911
     RUBY_SMART_CHAIN_TESTNET = 1912
     UPB_CRESCDI_TESTNET = 1918
+    SWELLCHAIN = 1923
+    SWELLCHAIN_TESTNET = 1924
+    ARVIX_TESTNET = 1927
     ONUS_CHAIN_TESTNET = 1945
+    MINATO = 1946
+    BIONIX_TESTNET = 1949
     D_CHAIN_MAINNET = 1951
     SELENDRA_NETWORK_TESTNET = 1953
     DEXILLA_TESTNET = 1954
@@ -636,7 +699,7 @@ class EthereumNetwork(Enum):
     PHALA_NETWORK = 2035
     KIWI_SUBNET = 2037
     SHRAPNEL_TESTNET = 2038
-    ALEPH_ZERO_TESTNET = 2039
+    ALEPH_ZERO = 2039
     VANAR_MAINNET = 2040
     NEUROWEB = 2043
     SHRAPNEL_SUBNET = 2044
@@ -644,6 +707,7 @@ class EthereumNetwork(Enum):
     STRATOS_TESTNET = 2047
     STRATOS = 2048
     MOVO_SMART_CHAIN_MAINNET = 2049
+    METACCES_MAINNET = 2071
     QUOKKACOIN_MAINNET = 2077
     ALTAIR = 2088
     ALGOL = 2089
@@ -664,6 +728,7 @@ class EthereumNetwork(Enum):
     FINDORA_TESTNET = 2153
     FINDORA_FORGE = 2154
     ANIMECHAIN_TESTNET = 2162
+    GAME7 = 2187
     SNAXCHAIN = 2192
     MOONSAMA_NETWORK = 2199
     ANTOFY_MAINNET = 2202
@@ -689,16 +754,20 @@ class EthereumNetwork(Enum):
     NEXIS_NETWORK_TESTNET = 2370
     BOMB_CHAIN_TESTNET = 2399
     TCG_VERSE_MAINNET = 2400
-    KARAK_MAINNET = 2410
+    K2_MAINNET = 2410
     XODEX = 2415
     KING_OF_LEGENDS_MAINNET = 2425
     POLYGON_ZKEVM_CARDONA_TESTNET = 2442
     HYBRID_CHAIN_NETWORK_TESTNET = 2458
     HYBRID_CHAIN_NETWORK_MAINNET = 2468
+    _6DEGREE_OF_OUTREACH = 2477
     UNICORN_ULTRA_NEBULAS_TESTNET = 2484
+    NOW_CHAIN_MAINNET = 2488
     KARAK_GOERLI = 2511
+    K2_TESTNET = 2512
     FRAXTAL_TESTNET = 2522
     INEVM_MAINNET = 2525
+    BAHAMUT_HORIZON = 2552
     KORTHO_MAINNET = 2559
     TECHPAY_MAINNET = 2569
     POCRNET = 2606
@@ -713,11 +782,15 @@ class EthereumNetwork(Enum):
     K_LAOS = 2718
     XR_SEPOLIA = 2730
     ELIZABETH_TESTNET = 2731
+    ABSTRACT = 2741
     NANON = 2748
     GM_NETWORK_MAINNET = 2777
     MORPH_HOLESKY = 2810
+    MORPH = 2818
+    HYPERAGI_MAINNET = 2868
     CHIPS_NETWORK = 2882
     BOBA_NETWORK_GOERLI_TESTNET = 2888
+    AARMA_MAINNET = 2889
     ELUX_CHAIN = 2907
     HYCHAIN = 2911
     XENON_CHAIN_TESTNET = 2941
@@ -740,10 +813,14 @@ class EthereumNetwork(Enum):
     DUBXCOIN_TESTNET = 3270
     DEBOUNCE_SUBNET_TESTNET = 3306
     ZCORE_TESTNET = 3331
+    ETHSTORAGE_L2_MAINNET = 3332
     ETHSTORAGE_TESTNET = 3333
     WEB3Q_GALILEO = 3334
-    ETHSTORAGE_MAINNET = 3335
+    ETHSTORAGE_L2_DEVNET = 3335
+    ETHSTORAGE_L2_TESTNET = 3336
+    ETHSTORAGE_DEVNET = 3337
     PEAQ = 3338
+    ETHSTORAGE_MAINNET = 3339
     PARIBU_NET_MAINNET = 3400
     EVOLVE_MAINNET = 3424
     SECURECHAIN_TESTNET = 3434
@@ -751,6 +828,7 @@ class EthereumNetwork(Enum):
     GTCSCAN = 3490
     PARIBU_NET_TESTNET = 3500
     JFIN_CHAIN = 3501
+    JZERO = 3502
     PANDOPROJECT_MAINNET = 3601
     PANDOPROJECT_TESTNET = 3602
     TYCOONCOIN = 3630
@@ -788,7 +866,8 @@ class EthereumNetwork(Enum):
     NAHMII_3_MAINNET = 4061
     NAHMII_3_TESTNET = 4062
     MUSTER_MAINNET = 4078
-    TOBE_CHAIN = 4080
+    TOBE_CHAIN_TESTNET = 4080
+    ZEROTH_MAINNET = 4088
     FASTEX_CHAIN_BAHAMUT_OASIS_TESTNET = 4090
     BITINDI_TESTNET = 4096
     BITINDI_MAINNET = 4099
@@ -796,6 +875,7 @@ class EthereumNetwork(Enum):
     HUMANS_AI_TESTNET = 4139
     TIPBOXCOIN_TESTNET = 4141
     CROSSFI_TESTNET = 4157
+    CROSSFI_MAINNET = 4158
     SX_ROLLUP = 4162
     PHI_NETWORK_V1 = 4181
     MERLIN_MAINNET = 4200
@@ -803,15 +883,21 @@ class EthereumNetwork(Enum):
     LISK_SEPOLIA_TESTNET = 4202
     NEXI_MAINNET = 4242
     NEXI_V2_MAINNET = 4243
+    LAIKA_TESTNET = 4269
+    ECHOS_CHAIN = 4321
     BOBAFUJI_TESTNET = 4328
     BEAM = 4337
     CREDIT_SMART_CHAIN_MAINNET = 4400
+    TESTNET_PIKA = 4422
     HTMLCOIN_MAINNET = 4444
     ORDERLY_SEPOLIA_TESTNET = 4460
     HYDRA_CHAIN = 4488
     EMONEY_NETWORK_TESTNET = 4544
+    EMONEY_NETWORK_MAINNET = 4545
     VERY_MAINNET = 4613
+    MST_CHAIN = 4646
     GOLD_CHAIN = 4653
+    APPCHAIN_TESTNET = 4661
     IOTEX_NETWORK_MAINNET = 4689
     IOTEX_NETWORK_TESTNET = 4690
     MEVERSE_CHAIN_TESTNET = 4759
@@ -828,7 +914,10 @@ class EthereumNetwork(Enum):
     TREASURENET_TESTNET = 5005
     ONIGIRI_TEST_SUBNET = 5039
     ONIGIRI_SUBNET = 5040
+    SKATE_MAINNET = 5050
     NOLLIE_SKATECHAIN_TESTNET = 5051
+    PIONEER_ZERO_CHAIN = 5080
+    PIONEER_CHAIN_TESTNET = 5090
     SYNDICATE_TESTNET = 5100
     SYNDICATE_FRAME_CHAIN = 5101
     SIC_TESTNET = 5102
@@ -837,6 +926,8 @@ class EthereumNetwork(Enum):
     SUPERLOYALTY_TESTNET = 5105
     AZRA_TESTNET = 5106
     HAM = 5112
+    CITREA_TESTNET = 5115
+    MECHAIN_TESTNET = 5151
     BAHAMUT = 5165
     SMART_LAYER_NETWORK = 5169
     TLCHAIN_NETWORK_MAINNET = 5177
@@ -846,12 +937,16 @@ class EthereumNetwork(Enum):
     UZMI_NETWORK_MAINNET = 5315
     OPTRUST_TESTNET = 5317
     ITX_TESTNET = 5321
+    SUPERSEED = 5330
     NETSBO = 5333
     TRITANIUM_TESTNET = 5353
     SETTLUS_TESTNET = 5372
+    SETTLUS_SEPOLIA_TESTNET = 5373
     EDEXA_MAINNET = 5424
     EGOCHAIN = 5439
+    POINTPAY_MAINNET = 5511
     VEX_EVM_TESTNET = 5522
+    DUCKCHAIN_MAINNET = 5545
     NAHMII_2_MAINNET = 5551
     NAHMII_2_TESTNET = 5553
     CHAIN_VERSE_MAINNET = 5555
@@ -866,26 +961,32 @@ class EthereumNetwork(Enum):
     SATOSHICHAIN_TESTNET = 5758
     TANGLE = 5845
     ONTOLOGY_TESTNET = 5851
+    CHANG_CHAIN_FOUNDATION_MAINNET = 5858
     WEGOCHAIN_RUBIDIUM_MAINNET = 5869
     BOUNCEBIT_TESTNET = 6000
     BOUNCEBIT_MAINNET = 6001
+    BC_HYPER_CHAIN_TESTNET = 6060
     TRES_TESTNET = 6065
     TRES_MAINNET = 6066
     CASCADIA_TESTNET = 6102
     UPTN_TESTNET = 6118
     UPTN = 6119
+    RAILS = 6278
     LAOS = 6283
     AURA_EUPHORIA_TESTNET = 6321
     AURA_MAINNET = 6322
     DIGIT_SOUL_SMART_CHAIN = 6363
     CONNEXT_SEPOLIA = 6398
     PEERPAY = 6502
+    FLAMMA_TESTNET = 6550
     SCOLCOIN_WEICHAIN_TESTNET = 6552
     FOX_TESTNET_NETWORK = 6565
     PIXIE_CHAIN_MAINNET = 6626
     LATEST_CHAIN_TESTNET = 6660
     CYBRIA_MAINNET = 6661
     CYBRIA_TESTNET = 6666
+    STORCHAIN = 6667
+    EDGE_MATRIX_CHAIN = 6678
     IRISHUB = 6688
     OX_CHAIN = 6699
     PAXB_MAINNET = 6701
@@ -894,6 +995,10 @@ class EthereumNetwork(Enum):
     RACE_MAINNET = 6805
     RACE_TESTNET = 6806
     POOLS_MAINNET = 6868
+    MTT_MAINNET = 6880
+    NIBIRU_MAINNET = 6900
+    XYL_TESTNET = 6934
+    LAIKA_MAINNET = 6942
     TOMB_CHAIN_MAINNET = 6969
     POLYSMARTCHAIN = 6999
     ZETACHAIN_MAINNET = 7000
@@ -903,8 +1008,13 @@ class EthereumNetwork(Enum):
     PLANQ_MAINNET = 7070
     PLANQ_ATLAS_TESTNET = 7077
     NUME = 7100
+    _0XL3 = 7117
     HELP_THE_HOMELESS = 7118
     BITROCK_MAINNET = 7171
+    EXSAT_MAINNET = 7200
+    NIBIRU_TESTNET_1 = 7210
+    NIBIRU_DEVNET_3 = 7222
+    ZEUS_TESTNET = 7244
     XPLA_VERSE = 7300
     KLYNTAR = 7331
     HORIZEN_EON_MAINNET = 7332
@@ -918,21 +1028,29 @@ class EthereumNetwork(Enum):
     THE_ROOT_NETWORK_PORCINI_TESTNET = 7672
     CANTO = 7700
     CANTO_TESNET = 7701
+    PHRON_TESTNET = 7744
     BITROCK_TESTNET = 7771
     GDCC_MAINNET = 7774
     GDCC_TESTNET = 7775
+    PANDASEA_MAINNET = 7776
     RISE_OF_THE_WARBOTS_TESTNET = 7777
     ORENIUM_MAINNET_PROTOCOL = 7778
     OPENEX_LONG_TESTNET = 7798
     MAALCHAIN_TESTNET = 7860
+    MAALCHAIN_V2 = 7862
+    MAALCHAIN_TESTNET_V2 = 7863
     HAZLOR_TESTNET = 7878
+    VEXON_TESTNET = 7879
     KINTO_MAINNET = 7887
     ARDENIUM_ATHENA = 7895
+    ARENA_Z = 7897
     DOT_BLOX = 7923
     MO_MAINNET = 7924
     DOS_CHAIN = 7979
     TELEPORT = 8000
     TELEPORT_TESTNET = 8001
+    POLYNOMIAL = 8008
+    ISUNCOIN_MAINNET = 8017
     MDGL_TESTNET = 8029
     BOAT_MAINNET = 8047
     KARAK_SEPOLIA = 8054
@@ -942,6 +1060,7 @@ class EthereumNetwork(Enum):
     BITCOIN_CHAIN = 8086
     E_DOLLAR = 8087
     STREAMUX_BLOCKCHAIN = 8098
+    ZENCHAIN = 8108
     QITMEER_NETWORK_TESTNET = 8131
     QITMEER_NETWORK_MIXNET = 8132
     QITMEER_NETWORK_PRIVNET = 8133
@@ -958,6 +1077,8 @@ class EthereumNetwork(Enum):
     LORENZO = 8329
     B3 = 8333
     DRACONES_FINANCIAL_SERVICES = 8387
+    ZENCHAIN_TESTNET = 8408
+    THAT_MAINNET = 8428
     BASE = 8453
     CHAKRA_TESTNET = 8545
     TOKI_NETWORK = 8654
@@ -998,9 +1119,13 @@ class EthereumNetwork(Enum):
     BERYLBIT_MAINNET = 9012
     NEXA_TESTNET_BLOCK = 9024
     NEXA_MAINNET_BLOCK = 9025
+    APEX_FUSION_NEXUS_MAINNET = 9069
+    APEX_FUSION_NEXUS_TESTNET = 9070
+    KPA_SMART_CHAIN_TESTNET = 9091
     GENESIS_COIN = 9100
     RINIA_TESTNET_OLD = 9170
     CODEFIN_MAINNET = 9223
+    GALACTICA_RETICULUM = 9302
     DOGCOIN_TESTNET = 9339
     OASYS_TESTNET = 9372
     DELA_SEPOLIA_TESTNET = 9393
@@ -1013,6 +1138,7 @@ class EthereumNetwork(Enum):
     BOBA_BNB_TESTNET = 9728
     MAINNETZ_TESTNET = 9768
     PEPENETWORK_MAINNET = 9779
+    TABI_TESTNETV2 = 9788
     TABI_TESTNET = 9789
     CARBON_EVM = 9790
     CARBON_EVM_TESTNET = 9792
@@ -1021,11 +1147,15 @@ class EthereumNetwork(Enum):
     IMPERIUM_MAINNET = 9819
     BINARYCHAIN_TESTNET = 9876
     DOGELAYER_MAINNET = 9888
+    ARENA_Z_TESTNET = 9897
     LARISSA_CHAIN = 9898
+    ZYTRON_LINEA_MAINNET = 9901
     ESPENTO_MAINNET = 9911
+    UXER_TESTNET_NETWORK = 9966
     MIND_SMART_CHAIN_TESTNET = 9977
     COMBO_MAINNET = 9980
     VOLLEY_MAINNET = 9981
+    MFEV_CHAIN_MAINNET = 9982
     AGUNG_NETWORK = 9990
     MIND_SMART_CHAIN_MAINNET = 9996
     ALTLAYER_TESTNET = 9997
@@ -1033,11 +1163,14 @@ class EthereumNetwork(Enum):
     MYOWN_TESTNET = 9999
     SMART_BITCOIN_CASH = 10000
     SMART_BITCOIN_CASH_TESTNET = 10001
+    WARDEN_TESTNET = 10010
     GON_CHAIN = 10024
+    AEON_CHAIN = 10025
     JAPAN_OPEN_CHAIN_TESTNET = 10081
     SJATSH = 10086
     METANOVA_VERSE = 10096
     BLOCKCHAIN_GENESIS_MAINNET = 10101
+    MONAD_TESTNET = 10143
     GNOSIS_CHIADO_TESTNET = 10200
     MAXXCHAIN_MAINNET = 10201
     GLSCAN = 10222
@@ -1052,9 +1185,12 @@ class EthereumNetwork(Enum):
     CRYPTOCOINPAY = 10823
     LAMINA1 = 10849
     LAMINA1_IDENTITY = 10850
+    GAMESWIFT_CHAIN_TESTNET = 10888
     QUADRANS_BLOCKCHAIN = 10946
     QUADRANS_BLOCKCHAIN_TESTNET = 10947
+    KB_CHAIN = 11000
     SHAPE_SEPOLIA_TESTNET = 11011
+    BOOL_NETWORK_BETA_MAINNET = 11100
     ASTRA = 11110
     WAGMI = 11111
     ASTRA_TESTNET = 11115
@@ -1071,6 +1207,7 @@ class EthereumNetwork(Enum):
     ARTELA_TESTNET = 11822
     SANTIMENT_INTELLIGENCE_NETWORK_DEPRECATED = 11888
     POLYGON_SUPERNET_ARIANEE = 11891
+    FUSE_TESTNET = 12001
     SATOSHICHAIN_MAINNET = 12009
     ATERNOS = 12020
     SINGULARITY_ZERO_TESTNET = 12051
@@ -1081,6 +1218,7 @@ class EthereumNetwork(Enum):
     L3X_PROTOCOL = 12324
     L3X_PROTOCOL_TESTNET = 12325
     STEP_TESTNET = 12345
+    GDPR_MAINNET = 12358
     RSS3_VSL_MAINNET = 12553
     RIKEZA_NETWORK_TESTNET = 12715
     PLAYDAPP_TESTNET = 12781
@@ -1089,16 +1227,21 @@ class EthereumNetwork(Enum):
     SPS = 13000
     CREDIT_SMART_CHAIN = 13308
     BEAM_TESTNET = 13337
+    CANNON_TESTNET = 13370
     IMMUTABLE_ZKEVM = 13371
     PHOENIX_MAINNET = 13381
     MASA = 13396
     IMMUTABLE_ZKEVM_TESTNET = 13473
     GRAVITY_ALPHA_TESTNET_SEPOLIA = 13505
     KRONOBIT_MAINNET = 13600
+    GAME7_TESTNET = 13746
     SUSONO = 13812
     SPS_TESTNET = 14000
+    ZEROTH_TESTNET = 14088
+    BITHARVEST_CHAIN_MAINNET = 14149
     EVOLVE_TESTNET = 14324
     VITRUVEO_TESTNET = 14333
+    VANA_MOKSHA_TESTNET = 14800
     VANA_SATORI_TESTNET = 14801
     HUMANODE_TESTNET_5_ISRAFEL = 14853
     IMMUTABLE_ZKEVM_DEVNET = 15003
@@ -1111,7 +1254,9 @@ class EthereumNetwork(Enum):
     METADOT_MAINNET = 16000
     METADOT_TESTNET = 16001
     DEFIVERSE_MAINNET = 16116
+    CYPHERIUM_MAINNET = 16166
     PLYR_PHI = 16180
+    EVENTUM_TESTNET = 16182
     INCENTIV_DEVNET = 16350
     PIVOTAL_SEPOLIA = 16481
     GENESYS_MAINNET = 16507
@@ -1136,26 +1281,36 @@ class EthereumNetwork(Enum):
     UNREAL_OLD = 18231
     UNREAL = 18233
     MXC_ZKEVM_MOONCHAIN = 18686
+    EXPCHAIN_TESTNET = 18880
     TITAN_TKX = 18888
     TITAN_TKX_TESTNET = 18889
     HOME_VERSE_MAINNET = 19011
     BLOCKX_ATLANTIS_TESTNET = 19077
+    LOCACHAIN_MAINNET = 19180
     BLOCKX_MAINNET = 19191
     DECENTRACONNECT_SOCIAL = 19224
+    SEC_TESTNET = 19515
+    SEC_MAINNET = 19516
     MAGNET_NETWORK = 19527
+    ZYTRON_LINEA_TESTNET = 19546
     LBRY_MAINNET = 19600
+    ASTRA_SEPOLIA = 19777
     BTCIX_NETWORK = 19845
     CAMELARK_MAINNET = 20001
     NIZA_CHAIN_MAINNET = 20041
     NIZA_CHAIN_TESTNET = 20073
+    MONAD_DEVNET = 20143
     CALLISTO_TESTNET = 20729
     P12_CHAIN = 20736
     JONO11_SUBNET = 20765
     C4EI = 21004
+    RIVEST_TESTNET = 21097
     ALL_ABOUT_HEALTHY = 21133
+    _1MONEY_NETWORK_MAINNET = 21210
     DCPAY_MAINNET = 21223
     DCPAY_TESTNET = 21224
     CENNZNET_AZALEA = 21337
+    LESTNET = 21363
     OMCHAIN_MAINNET = 21816
     BSL_MAINNET = 21912
     TAYCAN = 22023
@@ -1171,13 +1326,17 @@ class EthereumNetwork(Enum):
     DREYERX_TESTNET = 23452
     BLAST_TESTNET = 23888
     KYMTC_TESTNET = 24076
+    AMAUTI = 24116
     WEBCHAIN = 24484
     MINTME_COM_COIN = 24734
     LIQUIDLAYER_MAINNET = 25186
+    EVERCLEAR_MAINNET = 25327
     ALVEYCHAIN_TESTNET = 25839
     HAMMER_CHAIN_MAINNET = 25888
     BITKUB_CHAIN_TESTNET = 25925
     FERRUM_TESTNET = 26026
+    DUCATUSX_TESTNET = 26482
+    DUCATUSX = 26483
     HERTZ_NETWORK_MAINNET = 26600
     OASISCHAIN_MAINNET = 26863
     KLAOS_NOVA = 27181
@@ -1189,8 +1348,10 @@ class EthereumNetwork(Enum):
     BOBA_SEPOLIA = 28882
     HYCHAIN_TESTNET = 29112
     NEXA_METANET = 29223
+    NEXA_METATEST = 29225
     KAICHAIN_TESTNET = 29536
     MCH_VERSE_MAINNET = 29548
+    QCHAIN_MAINNET = 30000
     PIECE_TESTNET = 30067
     MIYOU_MAINNET = 30088
     CERIUM_TESTNET = 30103
@@ -1202,7 +1363,8 @@ class EthereumNetwork(Enum):
     CLOUDTX_TESTNET = 31224
     GOCHAIN_TESTNET = 31337
     EVOKE_TESTNET = 31414
-    FILECOIN_WALLABY_TESTNET = 31415
+    WIREX_PAY_MAINNET = 31415
+    MEZO_MATSNET_TESTNET = 31611
     XCHAIN_MAINNET = 31753
     XCHAIN_TESTNET = 31754
     W3GAMEZ_HOLESKY_TESTNET = 32001
@@ -1210,11 +1372,14 @@ class EthereumNetwork(Enum):
     BITGERT_MAINNET = 32520
     FUSION_MAINNET = 32659
     ZILLIQA_EVM = 32769
+    ZILLIQA_2_EVM_PROTO_MAINNET = 32770
     ZILLIQA_EVM_ISOLATED_SERVER = 32990
     ENTANGLE_MAINNET = 33033
     ZILLIQA_EVM_TESTNET = 33101
     ZILLIQA_2_EVM_PROTO_TESTNET = 33103
+    CURTIS = 33111
     ENTANGLE_TESTNET = 33133
+    APECHAIN = 33139
     CLOUDVERSE_SUBNET = 33210
     AVES_MAINNET = 33333
     ZILLIQA_EVM_DEVNET = 33385
@@ -1222,16 +1387,20 @@ class EthereumNetwork(Enum):
     ZILLIQA_2_EVM_DEVNET = 33469
     FUNKI = 33979
     MODE = 34443
+    ZEUS_MAINNET = 34504
     J2O_TARO = 35011
     Q_MAINNET = 35441
     Q_TESTNET = 35443
+    LENS_TESTNET = 37111
     CONNECTORMANAGER = 38400
     CONNECTORMANAGER_ROBIN = 38401
     PRM_MAINNET = 39656
     ENERGI_MAINNET = 39797
     OHO_MAINNET = 39815
+    DIV_CHAIN = 40000
     ALEPH_ZERO_EVM = 41455
     OPULENT_X_BETA = 41500
+    EDU_CHAIN = 41923
     PEGGLECOIN = 42069
     AGENTLAYER_TESTNET = 42072
     ARBITRUM_ONE = 42161
@@ -1240,6 +1409,8 @@ class EthereumNetwork(Enum):
     OASIS_EMERALD_TESTNET = 42261
     OASIS_EMERALD = 42262
     GOLDXCHAIN_MAINNET = 42355
+    ASSET_CHAIN_MAINNET = 42420
+    ASSET_CHAIN_TESTNET = 42421
     ZKFAIR_MAINNET = 42766
     ETHERLINK_MAINNET = 42793
     GESOTEN_VERSE_TESTNET = 42801
@@ -1249,6 +1420,7 @@ class EthereumNetwork(Enum):
     AVALANCHE_FUJI_TESTNET = 43113
     AVALANCHE_C_CHAIN = 43114
     BOBA_AVAX = 43288
+    FORMICARIUM = 43521
     ZKFAIR_TESTNET = 43851
     FRENCHAIN = 44444
     QUANTUM_NETWORK = 44445
@@ -1256,6 +1428,7 @@ class EthereumNetwork(Enum):
     AUTOBAHN_NETWORK = 45000
     SWAMPS_L2 = 45454
     DEELANCE_MAINNET = 45510
+    BLESSNET = 45513
     FUSION_TESTNET = 46688
     NEO_X_MAINNET = 47763
     REDEFI_LAYER_1 = 47803
@@ -1271,27 +1444,38 @@ class EthereumNetwork(Enum):
     YOOLDO_VERSE_MAINNET = 50005
     YOOLDO_VERSE_TESTNET = 50006
     GTON_TESTNET = 50021
+    SOPHON = 50104
+    REDDIO_DEVNET = 50341
+    ERBIE_MAINNET = 50888
     LUMOZ_TESTNET_ALPHA = 51178
     SARDIS_MAINNET = 51712
     ELECTRONEUM_MAINNET = 52014
+    CYTONIC_TESTNET = 52225
     DOID = 53277
     SUPERSEED_SEPOLIA_TESTNET = 53302
     DODOCHAIN_TESTNET = 53457
     DFK_CHAIN = 53935
+    OVERPROTOCOL_MAINNET = 54176
     HAQQ_CHAIN_TESTNET = 54211
     TORONET_TESTNET = 54321
     TITAN = 55004
+    TITAN_SEPOLIA = 55007
+    SUPERPOSITION = 55244
     PHOTON_AURORA_TESTNET = 55551
     REI_CHAIN_MAINNET = 55555
     REI_CHAIN_TESTNET = 55556
+    FLAMMA_MAINNET = 55614
     LAMBDA_CHAIN_MAINNET = 56026
     BOBA_BNB_MAINNET = 56288
     TESTNET_ZEROONE_SUBNET = 56400
     VELO_LABS_MAINNET = 56789
     DOID_TESTNET = 56797
     ROLLUX_TESTNET = 57000
+    SONIC_BLAZE_TESTNET = 57054
+    INK = 57073
     COINSEC_NETWORK = 57451
     SEPOLIA_PGN_PUBLIC_GOODS_NETWORK = 58008
+    LUMOZ_QUIDDITCH_TESTNET = 58680
     LINEA_GOERLI = 59140
     LINEA_SEPOLIA = 59141
     LINEA = 59144
@@ -1303,36 +1487,43 @@ class EthereumNetwork(Enum):
     THINKIUM_TESTNET_CHAIN_103 = 60103
     BOB = 60808
     ORANGE_CHAIN_MAINNET = 61022
+    TREASURE = 61166
     KAICHAIN = 61406
     AXELCHAIN_DEV_NET = 61800
     ETICA_MAINNET = 61803
     DOKEN_SUPER_CHAIN_MAINNET = 61916
     OPTOPIA_TESTNET = 62049
     OPTOPIA_MAINNET = 62050
+    TIKTRIX_TESTNET = 62092
     CITREA_DEVNET = 62298
     CELO_BAKLAVA_TESTNET = 62320
     MULTIVAC_MAINNET = 62621
     PLYR_TAU_TESTNET = 62831
     LAOS_SIGMA_TESTNET = 62850
-    ECREDITS_MAINNET = 63000
+    ESYNC_NETWORK_MAINNET = 63000
     ECREDITS_TESTNET = 63001
+    ESYNC_NETWORK_TESTNET = 63002
+    GEIST_MAINNET = 63157
     CRATD2C_TESTNET = 65349
     VECNO_MAINNET = 65357
     SCOLCOIN_MAINNET = 65450
     AUTOMATA_MAINNET = 65536
+    CREATOR_CHAIN_TESTNET = 66665
     JANUS_TESTNET = 66988
     SIRIUSNET = 67390
     COSMIC_CHAIN = 67588
     DM2_VERSE_MAINNET = 68770
+    DM2_VERSE_TESTNET = 68775
     CONDRIEU = 69420
     THINKIUM_MAINNET_CHAIN_0 = 70000
     THINKIUM_MAINNET_CHAIN_1 = 70001
     THINKIUM_MAINNET_CHAIN_2 = 70002
     THINKIUM_MAINNET_CHAIN_103 = 70103
     PROOF_OF_PLAY_APEX = 70700
+    PROOF_OF_PLAY_BOSS = 70701
     GUAPCOINX = 71111
+    WADZCHAIN_TESTNET = 71117
     POLYJUICE_TESTNET = 71393
-    GODWOKEN_TESTNET_V1 = 71401
     GODWOKEN_MAINNET = 71402
     CAGA_CRYPTO_ANKARA_TESTNET = 72778
     GROK_CHAIN_MAINNET = 72992
@@ -1341,6 +1532,7 @@ class EthereumNetwork(Enum):
     ENERGY_WEB_VOLTA_TESTNET = 73799
     MIXIN_VIRTUAL_MACHINE = 73927
     RESINCOIN_MAINNET = 75000
+    APPLAYER_TESTNET = 75338
     GEEK_VERSE_MAINNET = 75512
     GEEK_VERSE_TESTNET = 75513
     BORACHAIN_MAINNET = 77001
@@ -1357,6 +1549,7 @@ class EthereumNetwork(Enum):
     GOLD_SMART_CHAIN_TESTNET = 79879
     MUMBAI = 80001
     AMOY = 80002
+    POLYNOMIA_SEPOLIA = 80008
     BERACHAIN_BARTIO = 80084
     BERACHAIN_ARTIO = 80085
     HIZOCO_MAINNET = 80096
@@ -1374,20 +1567,27 @@ class EthereumNetwork(Enum):
     QUANTUM_CHAIN_MAINNET = 81720
     SMART_LAYER_NETWORK_TESTNET = 82459
     XPROTOCOL_TESTNET = 83144
+    ESA = 83278
     ZEDXION = 83872
     BASE_GOERLI_TESTNET = 84531
     BASE_SEPOLIA_TESTNET = 84532
+    O_CHAIN = 84841
     AERIE_NETWORK = 84886
+    GDPR_TESTNET = 85321
     CYBERTRUST = 85449
     NAUTILUS_PROTEUS_TESTNET = 88002
-    INOAI_NETWORK = 88559
+    INOAI = 88559
     ZKASINO_MAINNET = 88800
+    UNIT_ZERO_MAINNET = 88811
     UNIT_ZERO_TESTNET = 88817
     UNIT_ZERO_STAGENET = 88819
+    MATR1X_TESTNET = 88866
     CHILIZ_SCOVILLE_TESTNET = 88880
     CHILIZ_SPICY_TESTNET = 88882
     CHILIZ_CHAIN_MAINNET = 88888
+    UNITE = 88899
     F_XCORE_TESTNET_NETWORK = 90001
+    UBIT_SMARTCHAIN_MAINNET = 90002
     BEVERLY_HILLS = 90210
     CAMP_TESTNET = 90354
     NAUTILUS_TRITION_CHAIN = 91002
@@ -1396,13 +1596,23 @@ class EthereumNetwork(Enum):
     LAMBDA_TESTNET = 92001
     LIQUIDLAYER_TESTNET = 93572
     STRATOVM_TESTNET = 93747
+    SRICHAIN = 95432
+    LUX_TESTNET = 96368
+    LUX_MAINNET = 96369
     MANTIS_TESTNET_HEXAPOD = 96970
+    TETRON_TESTNET_SMART_CHAIN = 97053
+    TETRON_SMART_CHAIN = 97055
     BOBA_BNB_MAINNET_OLD = 97288
     SLINGSHOT_TESTNET = 97435
     GREEN_CHAIN_TESTNET = 97531
+    METABENZ_CHAIN = 97766
     OPTIMUSZ7_TESTNET = 97970
+    PLUME_DEVNET = 98864
+    PLUME_MAINNET = 98865
     EBI_CHAIN = 98881
+    SUPERPOSITION_TESTNET = 98985
     ELIBERTY_TESTNET = 99099
+    EDGE_MATRIX_CHAIN_TESTNET = 99876
     UB_SMART_CHAIN_TESTNET = 99998
     UB_SMART_CHAIN = 99999
     QUARKCHAIN_MAINNET_ROOT = 100000
@@ -1419,7 +1629,9 @@ class EthereumNetwork(Enum):
     QUARKCHAIN_L2_MAINNET = 100011
     DEPRECATED_CHI = 100100
     GLOBAL_TRUST_NETWORK = 101010
+    CREDITCOIN = 102030
     CREDITCOIN_TESTNET = 102031
+    CREDITCOIN_DEVNET = 102032
     CRYSTALEUM = 103090
     MASA_TESTNET = 103454
     KASPACLASSIC_MAINNET = 104566
@@ -1443,12 +1655,17 @@ class EthereumNetwork(Enum):
     METADAP_ENTERPRISE_TESTNET = 119139
     GEMCHAIN = 123321
     ADIL_DEVNET = 123456
+    MITOSIS_TESTNET = 124832
     ETHERLINK_TESTNET = 128123
     ODYSSEY_CHAIN_TESTNET = 131313
     ETND_CHAIN_MAINNETS = 131419
     FORM_TESTNET = 132902
     MAGAPE_TESTNET = 141319
+    BITHARVEST_CHAIN_TESTNET = 141491
     ICPLAZA_MAINNET = 142857
+    ODYSSEY_CHAIN_MAINNET = 153153
+    CRYPTOX = 158245
+    XCOIN = 158345
     PLAYFI_MAINNET = 161212
     ECLAT_MAINNET = 165279
     TAIKO_MAINNET = 167000
@@ -1458,14 +1675,19 @@ class EthereumNetwork(Enum):
     TAIKO_JOLNIR_L2 = 167007
     TAIKO_KATLA_L2 = 167008
     TAIKO_HEKLA_L2 = 167009
+    ZCHAINS = 168168
+    FAIR_TESTNET = 171000
+    WADZCHAIN_MAINNET = 171717
     CHRONICLE_LIT_PROTOCOL_TESTNET = 175177
     CHRONICLE_YELLOWSTONE_LIT_PROTOCOL_TESTNET = 175188
     BITICA_CHAIN_MAINNET = 188710
     CONDOR_TEST_NETWORK = 188881
     MIND_NETWORK_TESTNET = 192940
+    R0AR_CHAIN = 193939
     MAZZE_TESTNET = 199991
     XFAIR_AI_TESTNET = 200000
     MILKOMEDA_C1_TESTNET = 200101
+    ZOO_MAINNET = 200200
     MILKOMEDA_A1_TESTNET = 200202
     AKROMA = 200625
     BITLAYER_TESTNET = 200810
@@ -1474,13 +1696,19 @@ class EthereumNetwork(Enum):
     ALAYA_DEV_TESTNET = 201030
     MYTHICAL_CHAIN = 201804
     DECIMAL_SMART_CHAIN_TESTNET = 202020
+    DUCKCHAIN_TESTNET = 202105
+    BETHEL_SYDNEY = 202202
     X1_DEVNET = 202212
     YMTECH_BESU_TESTNET = 202401
+    BLOCKFIT = 202424
     JELLIE = 202624
     X1_NETWORK = 204005
     AURORIA_TESTNET = 205205
     GITAGI_ATLAS_TESTNET = 210049
+    SORIAN = 210209
+    SORIAN_TESTNET = 210210
     PLATON_MAINNET = 210425
+    LITENTRY = 212013
     MAS_MAINNET = 220315
     REAPCHAIN_MAINNET = 221230
     REAPCHAIN_TESTNET = 221231
@@ -1490,9 +1718,11 @@ class EthereumNetwork(Enum):
     TAF_ECO_CHAIN_MAINNET = 224168
     CONET_SEBOLIA_TESTNET = 224422
     CONET_HOLESKY = 224433
+    ABYSS_PROTOCOL = 229772
     HASHKEY_CHAIN_TESTNET_DISCARD = 230315
     HAYMO_TESTNET = 234666
     ORANGE_CHAIN_TESTNET = 240515
+    ANOMALY_ANDROMEDA_TESTNET = 241120
     ARTIS_SIGMA1 = 246529
     ARTIS_TESTNET_TAU1 = 246785
     SAAKURU_TESTNET = 247253
@@ -1500,14 +1730,18 @@ class EthereumNetwork(Enum):
     ECLAT_TESTNET = 262371
     GEAR_ZERO_NETWORK_TESTNET = 266256
     EGONCOIN_TESTNET = 271271
+    NXY_AREA_51 = 272247
+    NXY_OASIS = 272520
     SOCIAL_SMART_CHAIN_MAINNET = 281121
+    ATHENE_PARTHENON = 281123
     ZILLION_SEPOLIA_TESTNET = 282828
     CIPHEREM_TESTNET = 292003
-    ONE_WORLD_CHAIN = 309075
+    ONE_WORLD_CHAIN_MAINNET = 309075
     SAHARAAI_TESTNET = 313313
     FILECOIN_CALIBRATION_TESTNET = 314159
     PAREX_MAINNET = 322202
     BLOOM_GENESIS_TESTNET = 323213
+    CAMP_NETWORK_TESTNET_V2 = 325000
     WABA_CHAIN_TESTNET = 327126
     NAL_MAINNET = 328527
     TTCOIN_SMART_CHAIN_MAINNET = 330844
@@ -1527,6 +1761,7 @@ class EthereumNetwork(Enum):
     HAPCHAIN_TESTNET = 373737
     METAL_C_CHAIN = 381931
     METAL_TAHOE_C_CHAIN = 381932
+    CHEESECHAIN = 383353
     TIPBOXCOIN_MAINNET = 404040
     AIE_TESTNET = 413413
     KEKCHAIN = 420420
@@ -1557,14 +1792,19 @@ class EthereumNetwork(Enum):
     SCROLL_PRE_ALPHA_TESTNET = 534354
     SHINARIUM_BETA = 534849
     BEANECO_SMARTCHAIN = 535037
+    OVERPROTOCOL_TESTNET = 541764
+    ZERO_NETWORK = 543210
     ONE_WORLD_CHAIN_TESTNET = 552981
     PENTAGON_TESTNET = 555555
     ECLIPSE_TESTNET = 555666
+    DUSTBOY_IOT = 555888
+    DJT_TESTNET = 621847
     HYPRA_MAINNET = 622277
     ATLAS = 622463
+    POLTER_TESTNET = 631571
     BEAR_NETWORK_CHAIN_MAINNET = 641230
     ALL_MAINNET = 651940
-    OPEN_CAMPUS_CODEX = 656476
+    EDU_CHAIN_TESTNET = 656476
     XAI_MAINNET = 660279
     VISION_VPIONEER_TEST_CHAIN = 666666
     HELA_OFFICIAL_RUNTIME_TESTNET = 666888
@@ -1573,35 +1813,47 @@ class EthereumNetwork(Enum):
     TILTYARD_MAINNET_SUBNET = 710420
     SEI_DEVNET = 713715
     ERAM_MAINNET = 721529
+    TIXCHAIN_TESTNET = 723107
     HEMI_SEPOLIA = 743111
     BEAR_NETWORK_CHAIN_TESTNET = 751230
+    TERNOA_TESTNET = 752024
+    PAYSCAN_CHAIN = 756689
     MIEXS_SMARTCHAIN = 761412
+    INK_SEPOLIA = 763373
     LAMINA1_TESTNET = 764984
     LAMINA1_IDENTITY_TESTNET = 767368
     MODULARIUM = 776877
+    ZEBRO_SMART_CHAIN = 786786
     OCTASPACE = 800001
     BIZ_SMART_CHAIN_TESTNET = 808080
     BOB_SEPOLIA = 808813
     ZKLINK_NOVA_MAINNET = 810180
     ZKLINK_NOVA_SEPOLIA_TESTNET = 810181
     ZKLINK_NOVA_GOERLI_TESTNET = 810182
+    SG_VERSE_MAINNET = 812397
     TSC_TESTNET = 820522
     CURVE_MAINNET = 827431
     PRM_TESTNET = 839320
+    EXSAT_TESTNET = 839999
+    RUNEVM_TESTNET = 840000
     _4GOODNETWORK = 846000
     DODAO = 855456
     BLOCX_MAINNET = 879151
     REXX_MAINNET = 888882
     VISION_MAINNET = 888888
+    UNITE_TESTNET = 888991
     POSICHAIN_MAINNET_SHARD_0 = 900000
     POSICHAIN_TESTNET_SHARD_0 = 910000
+    ODYSSEY_TESTNET = 911867
     ASTRIA_EVM_DUSKNET = 912559
+    OWSHEN_MAINNET = 918273
     POSICHAIN_DEVNET_SHARD_0 = 920000
     POSICHAIN_DEVNET_SHARD_1 = 920001
     FNCY_TESTNET = 923018
     JONO12_SUBNET = 955081
     ELUVIO_CONTENT_FABRIC = 955305
     TREASURE_RUBY = 978657
+    TREASURE_TOPAZ = 978658
     FORMA = 984122
     FORMA_SKETCHPAD = 984123
     ECROX_CHAIN_MAINNET = 988207
@@ -1609,6 +1861,7 @@ class EthereumNetwork(Enum):
     AMCHAIN = 999999
     NETMIND_CHAIN_TESTNET = 1100789
     TILTYARD_SUBNET = 1127469
+    _1MONEY_NETWORK_TESTNET = 1212101
     SHARECLE_MAINNET = 1234567
     ZKATANA = 1261120
     ETHO_PROTOCOL = 1313114
@@ -1617,18 +1870,22 @@ class EthereumNetwork(Enum):
     KILN = 1337802
     ZHEJIANG = 1337803
     AUTOMATA_TESTNET = 1398243
+    AUTOMATA_ORBIT_TESTNET = 1398244
+    WATERFALL_9_TEST_NETWORK = 1501869
     PLAYFI_ALBIREO_TESTNET = 1612127
     XTERIO_TESTNET = 1637450
     TURKEY_DEMO_DEV = 1731313
+    XPHERE_TESTNET = 1998991
     DEBANK_TESTNET = 2021398
     PLIAN_MAINNET_MAIN = 2099156
     PLATON_DEV_TESTNET_DEPRECATED = 2203181
     PLATON_DEV_TESTNET2 = 2206132
+    COINWEB_BNB_SHARD = 2222222
     DPU_CHAIN = 2611555
-    XTERIO_CHAIN = 2702128
+    XTERIO_CHAIN_ETH = 2702128
     SAHARAAI_NETWORK = 3132023
     FILECOIN_BUTTERFLY_TESTNET = 3141592
-    FUNKI_SEPOLIA_SANDBOX = 3397901
+    FUNKI_SEPOLIA_TESTNET = 3397901
     MANTA_PACIFIC_TESTNET = 3441005
     MANTA_PACIFIC_SEPOLIA_TESTNET = 3441006
     ALTLAYER_ZERO_GAS_NETWORK = 4000003
@@ -1640,11 +1897,13 @@ class EthereumNetwork(Enum):
     MOONCHAIN_GENEVA_TESTNET = 5167004
     ELECTRONEUM_TESTNET = 5201420
     REACTIVE_KOPLI = 5318008
+    POINTPAY_TESTNET = 5511555
     IMVERSED_MAINNET = 5555555
     IMVERSED_TESTNET = 5555558
     ASTAR_ZKYOTO = 6038361
     SAFE_ANWANG_MAINNET = 6666665
     SAFE_ANWANG_TESTNET = 6666666
+    COTI_TESTNET = 7082400
     SAAKURU_MAINNET = 7225878
     OPENVESSEL = 7355310
     QL1_TESTNET = 7668378
@@ -1659,14 +1918,21 @@ class EthereumNetwork(Enum):
     QUARIX = 8888888
     XCAP = 9322252
     MILVINE = 9322253
+    FLUENCE = 9999999
+    SPOTLIGHT = 10058111
+    SPOTLIGHT_SEPOLIA_TESTNET = 10058112
     PLIAN_TESTNET_SUBCHAIN_1 = 10067275
     SOVERUN_MAINNET = 10101010
     ALIENX_MAINNET = 10241024
     ALIENX_HAL_TESTNET = 10241025
+    BLESSNET_SEPOLIA = 11145513
     SEPOLIA = 11155111
     OP_SEPOLIA_TESTNET = 11155420
+    R0AR_TESTNET = 11166111
+    LUMMIO_NETWORK = 12020498
+    MEMENTO_TESTNET = 12052024
     NEOX_TESTNET_T3 = 12227331
-    NEOX_TESTNET_T4 = 12227332
+    NEO_X_TESTNET_T4 = 12227332
     COTI_DEVNET = 13068200
     PEPCHAIN_CHURCHILL = 13371337
     ANDUSCHAIN_MAINNET = 14288640
@@ -1685,17 +1951,28 @@ class EthereumNetwork(Enum):
     DBK_CHAIN = 20240603
     SWAN_PROXIMA_TESTNET = 20241133
     HOKUM_TESTNET = 20482050
+    CORN_MAIZENET = 21000000
+    CORN_TESTNET = 21000001
     EXCELON_MAINNET = 22052002
+    XMTP = 24132016
+    _6DEGREE_OF_OUTREACH_TESTNET = 24772477
     EXCOINCIAL_CHAIN_VOLTA_TESTNET = 27082017
     EXCOINCIAL_CHAIN_MAINNET = 27082022
     ANCIENT8_TESTNET = 28122024
     AUXILIUM_NETWORK_MAINNET = 28945486
     FLACHAIN_MAINNET = 29032022
     FILECOIN_LOCAL_TESTNET = 31415926
+    TOLIMAN_SUAVE_TESTNET = 33626250
+    XONE_TESTNET = 33772211
+    CITRONUS_CITRO = 34949059
     JOYS_DIGITAL_MAINNET = 35855456
     SKALE_NEBULA_HUB_TESTNET = 37084624
     KINGDOM_CHAIN = 39916801
     MAISTESTSUBNET = 43214913
+    STAVANGER_PUBLIC_TESTNET = 50591822
+    DEVIANT_TOKEN_BLOCKCHAIN = 52027071
+    DEVIANT_TOKEN_BLOCKCHAIN_TESTNET = 52027080
+    FLUENCE_TESTNET = 52164803
     AQUACHAIN = 61717561
     AUTONITY_BAKERLOO_THAMES_TESTNET = 65010000
     AUTONITY_BAKERLOO_BARADA_TESTNET = 65010001
@@ -1710,8 +1987,11 @@ class EthereumNetwork(Enum):
     SX_TORONTO_ROLLUP = 79479957
     BACKSTOP_TESTNET = 88558801
     T_E_A_M_BLOCKCHAIN = 88888888
+    REYA_CRONOS = 89346162
     POLYGON_BLACKBERRY = 94204209
+    CHADCHAIN = 97912060
     JOYS_DIGITAL_TESTNET = 99415706
+    ETHOS = 100000000
     ORAICHAIN_MAINNET = 108160679
     CYBER_TESTNET = 111557560
     OP_CELESTIA_RASPBERRY = 123420111
@@ -1719,25 +1999,31 @@ class EthereumNetwork(Enum):
     BLAST_SEPOLIA_TESTNET = 168587773
     GATHER_MAINNET_NETWORK = 192837465
     KANAZAWA = 222000222
+    XMTP_SEPOLIA = 241320161
     NEON_EVM_DEVNET = 245022926
     NEON_EVM_DEVNET_ROLLUP = 245022929
     NEON_EVM_MAINNET = 245022934
     NEON_EVM_TESTNET = 245022940
+    FLAME = 253368190
     RAZOR_SKALE_CHAIN = 278611351
     ONELEDGER_MAINNET = 311752642
     NAL_SEPOLIA_TESTNET = 328527624
     MELD = 333000333
     DEPRECATED_SKALE_CALYPSO_HUB_TESTNET = 344106930
     GATHER_TESTNET_NETWORK = 356256156
+    WESTEND_ASSET_HUB = 420420421
     DEPRECATED_SKALE_EUROPA_HUB_TESTNET = 476158412
     SKOPJE_TESTNET = 476462898
     GATHER_DEVNET_NETWORK = 486217935
     DEPRECATED_SKALE_NEBULA_HUB_TESTNET = 503129905
+    SOPHON_TESTNET = 531050104
     DEGEN_CHAIN = 666666666
+    TRON_MAINNET = 728126428
     ANCIENT8 = 888888888
     PTCESCAN_TESTNET = 889910245
     PTCESCAN_MAINNET = 889910246
     SKALE_CALYPSO_HUB_TESTNET = 974399131
+    LUMIA_MAINNET = 994873017
     ZORA_SEPOLIA_TESTNET = 999999999
     SKALE_TITAN_HUB_TESTNET = 1020352220
     IPOS_NETWORK = 1122334455
@@ -1756,6 +2042,7 @@ class EthereumNetwork(Enum):
     GPT_MAINNET = 1511670449
     DEPRECATED_SKALE_TITAN_HUB_TESTNET = 1517929550
     SKALE_CALYPSO_HUB = 1564830818
+    HAUST_TESTNET = 1570754601
     HARMONY_MAINNET_SHARD_0 = 1666600000
     HARMONY_MAINNET_SHARD_1 = 1666600001
     HARMONY_MAINNET_SHARD_2 = 1666600002
@@ -1764,28 +2051,37 @@ class EthereumNetwork(Enum):
     HARMONY_TESTNET_SHARD_1 = 1666700001
     HARMONY_DEVNET_SHARD_0 = 1666900000
     HARMONY_DEVNET_SHARD_1 = 1666900001
-    KAKAROT_SEPOLIA = 1802203764
+    SILICON_ZKEVM_SEPOLIA_TESTNET = 1722641160
     GEMUCHAIN_TESTNET = 1903648807
     RARI_CHAIN_TESTNET = 1918988905
+    LUMIA_TESTNET = 1952959480
     DATAHOPPER = 2021121117
     SKALE_EUROPA_HUB = 2046399126
+    ACCUMULATE_KERMIT = 2478899481
     ANCIENT8_TESTNET_DEPRECATED = 2863311531
     PIRL = 3125659152
     ONELEDGER_TESTNET_FRANKENSTEIN = 4216137055
+    MEKONG = 7078815900
     PALM_TESTNET = 11297108099
     PALM = 11297108109
     GITSWARM_TEST_NETWORK = 28872323069
     XAI_TESTNET_V2 = 37714555429
+    DERIW_DEVNET = 44474237230
     ARBITRUM_BLUEBERRY = 88153591557
-    KAKAROT_SEPOLIA_DEPRECATED = 107107114116
     ALPHABET_MAINNET = 111222333444
+    THANOS_SEPOLIA = 111551119090
     FLUENCE_STAGE = 123420000220
+    PIN = 123420000558
+    SIVO_DEFI_TESTNET = 123420000586
+    VOLMEX = 123420000588
     NTITY_MAINNET = 197710212030
     HARADEV_TESTNET = 197710212031
     GM_NETWORK_TESTNET = 202402181627
     ZENIQ = 383414847825
     PDC_MAINNET = 666301171999
     MOLEREUM_NETWORK = 6022140761023
+    FLAME_TESTNET = 16604737732183
+    KAKAROT_STARKNET_SEPOLIA = 920637907288165
     DCHAIN_TESTNET = 2713017997578000
     DCHAIN = 2716446429837000
 
