@@ -26,7 +26,7 @@ user_operation_mock: Dict[str, Any] = {
         "userOperation": {
             "sender": "0xB0B5c0578Aa134b0496a6C0e51A7aae47C522861",
             "nonce": "0x0",
-            "initCode": "0x4e1dcf7ad4e460cfd30791ccc4f9c8a4f820ec671688f0b900000000000000000000000029fcb43b46531bca003ddc8fcb67ffe91900c7620000000000000000000000000000000000000000000000000000000000000060000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001e4b63e800d000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000000010000000000000000000000008ecd4ec46d4d2a6b64fe960b3d64e8b94b2234eb0000000000000000000000000000000000000000000000000000000000000140000000000000000000000000a581c4a4db7175302464ff3c06380bc3270b403700000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000005ac255889882acd3da2aa939679e3f3d4cea221e00000000000000000000000000000000000000000000000000000000000000648d0dc49f00000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000001000000000000000000000000a581c4a4db7175302464ff3c06380bc3270b40370000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
+            "initCode": "0x4e1DCf7AD4e460CfD30791CCC4F9c8a4f820ec671688f0b900000000000000000000000029fcb43b46531bca003ddc8fcb67ffe91900c7620000000000000000000000000000000000000000000000000000000000000060000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001e4b63e800d000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000000010000000000000000000000008ecd4ec46d4d2a6b64fe960b3d64e8b94b2234eb0000000000000000000000000000000000000000000000000000000000000140000000000000000000000000a581c4a4db7175302464ff3c06380bc3270b403700000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000010000000000000000000000005ac255889882acd3da2aa939679e3f3d4cea221e00000000000000000000000000000000000000000000000000000000000000648d0dc49f00000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000001000000000000000000000000a581c4a4db7175302464ff3c06380bc3270b40370000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
             "callData": "0x7bb3742800000000000000000000000002270bd144e70ce6963ba02f575776a16184e1e600000000000000000000000000000000000000000000000000005af3107a4000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
             "callGasLimit": "0x6000",
             "verificationGasLimit": "0x76766",
@@ -58,7 +58,7 @@ user_operation_receipt_mock: Dict[str, Any] = {
         "reason": "",
         "logs": [
             {
-                "address": "0xb0b5c0578aa134b0496a6c0e51a7aae47c522861",
+                "address": "0xB0B5c0578Aa134b0496a6C0e51A7aae47C522861",
                 "topics": [
                     "0xecdf3a3effea5783a3c4c2140e677577666428d44ed9d474a0b3a4c9943f8440",
                     "0x000000000000000000000000a581c4a4db7175302464ff3c06380bc3270b4037",
@@ -72,7 +72,7 @@ user_operation_receipt_mock: Dict[str, Any] = {
                 "removed": False,
             },
             {
-                "address": "0xb0b5c0578aa134b0496a6c0e51a7aae47c522861",
+                "address": "0xB0B5c0578Aa134b0496a6C0e51A7aae47C522861",
                 "topics": [
                     "0x141df868a6331af528e38c83b7aa03edc19be66e37ae67f9285bf4f8e3c6a1a8",
                     "0x0000000000000000000000004e1dcf7ad4e460cfd30791ccc4f9c8a4f820ec67",
@@ -86,7 +86,7 @@ user_operation_receipt_mock: Dict[str, Any] = {
                 "removed": False,
             },
             {
-                "address": "0x4e1dcf7ad4e460cfd30791ccc4f9c8a4f820ec67",
+                "address": "0x4e1DCf7AD4e460CfD30791CCC4F9c8a4f820ec67",
                 "topics": [
                     "0x4f51faf6c4561ff95f067657e43439f0f856d97c04d9ec9070a6199ad418e235",
                     "0x000000000000000000000000b0b5c0578aa134b0496a6c0e51a7aae47c522861",
@@ -100,7 +100,7 @@ user_operation_receipt_mock: Dict[str, Any] = {
                 "removed": False,
             },
             {
-                "address": "0x5ff137d4b0fdcd49dca30c7cf57e578a026d2789",
+                "address": "0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789",
                 "topics": [
                     "0xd51a9c61267aa6196961883ecf5ff2da6619c37dac0fa92122513fb32c032d2d",
                     "0x39b3e2171c04539d9b3f848d04364dfaa42cc0b412ff65ce2a85c566cf8bf281",
@@ -115,7 +115,7 @@ user_operation_receipt_mock: Dict[str, Any] = {
                 "removed": False,
             },
             {
-                "address": "0xb0b5c0578aa134b0496a6c0e51a7aae47c522861",
+                "address": "0xB0B5c0578Aa134b0496a6C0e51A7aae47C522861",
                 "topics": [
                     "0xb648d3644f584ed1c2232d53c46d87e693586486ad0d1175f8656013110b714e"
                 ],
@@ -128,7 +128,7 @@ user_operation_receipt_mock: Dict[str, Any] = {
                 "removed": False,
             },
             {
-                "address": "0x5ff137d4b0fdcd49dca30c7cf57e578a026d2789",
+                "address": "0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789",
                 "topics": [
                     "0x2da466a7b24304f47e87fa2e1e5a81b9831ce54fec19055ce277ca2f39ba42c4",
                     "0x000000000000000000000000b0b5c0578aa134b0496a6c0e51a7aae47c522861",
@@ -142,7 +142,7 @@ user_operation_receipt_mock: Dict[str, Any] = {
                 "removed": False,
             },
             {
-                "address": "0xb0b5c0578aa134b0496a6c0e51a7aae47c522861",
+                "address": "0xB0B5c0578Aa134b0496a6C0e51A7aae47C522861",
                 "topics": [
                     "0x6895c13664aa4f67288b25d7a21d7aaa34916e355fb9b6fae0a139a9085becb8",
                     "0x000000000000000000000000a581c4a4db7175302464ff3c06380bc3270b4037",
@@ -156,7 +156,7 @@ user_operation_receipt_mock: Dict[str, Any] = {
                 "removed": False,
             },
             {
-                "address": "0x5ff137d4b0fdcd49dca30c7cf57e578a026d2789",
+                "address": "0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789",
                 "topics": [
                     "0xbb47ee3e183a558b1a2ff0874b079f3fc5478b7454eacf2bfc5af2ff5878f972"
                 ],
@@ -169,7 +169,7 @@ user_operation_receipt_mock: Dict[str, Any] = {
                 "removed": False,
             },
             {
-                "address": "0xb0b5c0578aa134b0496a6c0e51a7aae47c522861",
+                "address": "0xB0B5c0578Aa134b0496a6C0e51A7aae47C522861",
                 "topics": [
                     "0xb648d3644f584ed1c2232d53c46d87e693586486ad0d1175f8656013110b714e"
                 ],
@@ -182,7 +182,7 @@ user_operation_receipt_mock: Dict[str, Any] = {
                 "removed": False,
             },
             {
-                "address": "0xb0b5c0578aa134b0496a6c0e51a7aae47c522861",
+                "address": "0xB0B5c0578Aa134b0496a6C0e51A7aae47C522861",
                 "topics": [
                     "0x6895c13664aa4f67288b25d7a21d7aaa34916e355fb9b6fae0a139a9085becb8",
                     "0x000000000000000000000000a581c4a4db7175302464ff3c06380bc3270b4037",
@@ -196,7 +196,7 @@ user_operation_receipt_mock: Dict[str, Any] = {
                 "removed": False,
             },
             {
-                "address": "0x5ff137d4b0fdcd49dca30c7cf57e578a026d2789",
+                "address": "0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789",
                 "topics": [
                     "0x49628fd1471006c1482da88028e9ce4dbb080b815c9b0344d39e5a8e6ec1419f",
                     "0x39b3e2171c04539d9b3f848d04364dfaa42cc0b412ff65ce2a85c566cf8bf281",
@@ -218,13 +218,13 @@ user_operation_receipt_mock: Dict[str, Any] = {
             "blockHash": "0xcc466b284f4030ee3f5941a2c8e36892262bf583611c902fe5558a595af47e13",
             "blockNumber": "0x50b0da",
             "from": "0xd53eb5203e367bbdd4f72338938224881fc501ab",
-            "to": "0x5ff137d4b0fdcd49dca30c7cf57e578a026d2789",
+            "to": "0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789",
             "cumulativeGasUsed": "0xd2a354",
             "gasUsed": "0x67c20",
             "contractAddress": None,
             "logs": [
                 {
-                    "address": "0xb0b5c0578aa134b0496a6c0e51a7aae47c522861",
+                    "address": "0xB0B5c0578Aa134b0496a6C0e51A7aae47C522861",
                     "topics": [
                         "0xecdf3a3effea5783a3c4c2140e677577666428d44ed9d474a0b3a4c9943f8440",
                         "0x000000000000000000000000a581c4a4db7175302464ff3c06380bc3270b4037",
@@ -238,7 +238,7 @@ user_operation_receipt_mock: Dict[str, Any] = {
                     "removed": False,
                 },
                 {
-                    "address": "0xb0b5c0578aa134b0496a6c0e51a7aae47c522861",
+                    "address": "0xB0B5c0578Aa134b0496a6C0e51A7aae47C522861",
                     "topics": [
                         "0x141df868a6331af528e38c83b7aa03edc19be66e37ae67f9285bf4f8e3c6a1a8",
                         "0x0000000000000000000000004e1dcf7ad4e460cfd30791ccc4f9c8a4f820ec67",
@@ -252,7 +252,7 @@ user_operation_receipt_mock: Dict[str, Any] = {
                     "removed": False,
                 },
                 {
-                    "address": "0x4e1dcf7ad4e460cfd30791ccc4f9c8a4f820ec67",
+                    "address": "0x4e1DCf7AD4e460CfD30791CCC4F9c8a4f820ec67",
                     "topics": [
                         "0x4f51faf6c4561ff95f067657e43439f0f856d97c04d9ec9070a6199ad418e235",
                         "0x000000000000000000000000b0b5c0578aa134b0496a6c0e51a7aae47c522861",
@@ -266,7 +266,7 @@ user_operation_receipt_mock: Dict[str, Any] = {
                     "removed": False,
                 },
                 {
-                    "address": "0x5ff137d4b0fdcd49dca30c7cf57e578a026d2789",
+                    "address": "0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789",
                     "topics": [
                         "0xd51a9c61267aa6196961883ecf5ff2da6619c37dac0fa92122513fb32c032d2d",
                         "0x39b3e2171c04539d9b3f848d04364dfaa42cc0b412ff65ce2a85c566cf8bf281",
@@ -281,7 +281,7 @@ user_operation_receipt_mock: Dict[str, Any] = {
                     "removed": False,
                 },
                 {
-                    "address": "0xb0b5c0578aa134b0496a6c0e51a7aae47c522861",
+                    "address": "0xB0B5c0578Aa134b0496a6C0e51A7aae47C522861",
                     "topics": [
                         "0xb648d3644f584ed1c2232d53c46d87e693586486ad0d1175f8656013110b714e"
                     ],
@@ -294,7 +294,7 @@ user_operation_receipt_mock: Dict[str, Any] = {
                     "removed": False,
                 },
                 {
-                    "address": "0x5ff137d4b0fdcd49dca30c7cf57e578a026d2789",
+                    "address": "0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789",
                     "topics": [
                         "0x2da466a7b24304f47e87fa2e1e5a81b9831ce54fec19055ce277ca2f39ba42c4",
                         "0x000000000000000000000000b0b5c0578aa134b0496a6c0e51a7aae47c522861",
@@ -308,7 +308,7 @@ user_operation_receipt_mock: Dict[str, Any] = {
                     "removed": False,
                 },
                 {
-                    "address": "0xb0b5c0578aa134b0496a6c0e51a7aae47c522861",
+                    "address": "0xB0B5c0578Aa134b0496a6C0e51A7aae47C522861",
                     "topics": [
                         "0x6895c13664aa4f67288b25d7a21d7aaa34916e355fb9b6fae0a139a9085becb8",
                         "0x000000000000000000000000a581c4a4db7175302464ff3c06380bc3270b4037",
@@ -322,7 +322,7 @@ user_operation_receipt_mock: Dict[str, Any] = {
                     "removed": False,
                 },
                 {
-                    "address": "0x5ff137d4b0fdcd49dca30c7cf57e578a026d2789",
+                    "address": "0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789",
                     "topics": [
                         "0xbb47ee3e183a558b1a2ff0874b079f3fc5478b7454eacf2bfc5af2ff5878f972"
                     ],
@@ -335,7 +335,7 @@ user_operation_receipt_mock: Dict[str, Any] = {
                     "removed": False,
                 },
                 {
-                    "address": "0xb0b5c0578aa134b0496a6c0e51a7aae47c522861",
+                    "address": "0xB0B5c0578Aa134b0496a6C0e51A7aae47C522861",
                     "topics": [
                         "0xb648d3644f584ed1c2232d53c46d87e693586486ad0d1175f8656013110b714e"
                     ],
@@ -348,7 +348,7 @@ user_operation_receipt_mock: Dict[str, Any] = {
                     "removed": False,
                 },
                 {
-                    "address": "0xb0b5c0578aa134b0496a6c0e51a7aae47c522861",
+                    "address": "0xB0B5c0578Aa134b0496a6C0e51A7aae47C522861",
                     "topics": [
                         "0x6895c13664aa4f67288b25d7a21d7aaa34916e355fb9b6fae0a139a9085becb8",
                         "0x000000000000000000000000a581c4a4db7175302464ff3c06380bc3270b4037",
@@ -362,7 +362,7 @@ user_operation_receipt_mock: Dict[str, Any] = {
                     "removed": False,
                 },
                 {
-                    "address": "0x5ff137d4b0fdcd49dca30c7cf57e578a026d2789",
+                    "address": "0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789",
                     "topics": [
                         "0x49628fd1471006c1482da88028e9ce4dbb080b815c9b0344d39e5a8e6ec1419f",
                         "0x39b3e2171c04539d9b3f848d04364dfaa42cc0b412ff65ce2a85c566cf8bf281",
@@ -408,7 +408,7 @@ user_operation_with_valid_dates_mock = {
         "userOperation": {
             "sender": "0xb17E6F96085b443389E30eEAc243DA37A71cbd42",
             "nonce": "0x0",
-            "initCode": "0x4e1dcf7ad4e460cfd30791ccc4f9c8a4f820ec671688f0b900000000000000000000000029fcb43b46531bca003ddc8fcb67ffe91900c76200000000000000000000000000000000000000000000000000000000000000600000000000000000000000000000000000000000000000000000000000069e3500000000000000000000000000000000000000000000000000000000000001e4b63e800d000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000000010000000000000000000000008ecd4ec46d4d2a6b64fe960b3d64e8b94b2234eb0000000000000000000000000000000000000000000000000000000000000140000000000000000000000000a581c4a4db7175302464ff3c06380bc3270b40370000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000d725e11588f040d86c4c49d8236e32a5868549f000000000000000000000000000000000000000000000000000000000000000648d0dc49f00000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000001000000000000000000000000a581c4a4db7175302464ff3c06380bc3270b40370000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
+            "initCode": "0x4e1DCf7AD4e460CfD30791CCC4F9c8a4f820ec671688f0b900000000000000000000000029fcb43b46531bca003ddc8fcb67ffe91900c76200000000000000000000000000000000000000000000000000000000000000600000000000000000000000000000000000000000000000000000000000069e3500000000000000000000000000000000000000000000000000000000000001e4b63e800d000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000000010000000000000000000000008ecd4ec46d4d2a6b64fe960b3d64e8b94b2234eb0000000000000000000000000000000000000000000000000000000000000140000000000000000000000000a581c4a4db7175302464ff3c06380bc3270b40370000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001000000000000000000000000d725e11588f040d86c4c49d8236e32a5868549f000000000000000000000000000000000000000000000000000000000000000648d0dc49f00000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000001000000000000000000000000a581c4a4db7175302464ff3c06380bc3270b40370000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
             "callData": "0x7bb3742800000000000000000000000038869bf66a61cf6bdb996a6ae40d5853fd43b52600000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000080000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000001848d80ff0a00000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000132001c7d4b196cb0c7b01d743fbc6116a902379c723800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000044a9059cbb000000000000000000000000b17e6f96085b443389e30eeac243da37a71cbd4200000000000000000000000000000000000000000000000000000000000186a0001c7d4b196cb0c7b01d743fbc6116a902379c723800000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000044a9059cbb000000000000000000000000b17e6f96085b443389e30eeac243da37a71cbd4200000000000000000000000000000000000000000000000000000000000186a0000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
             "callGasLimit": "0x1d006",
             "verificationGasLimit": "0x840a0",
@@ -444,7 +444,7 @@ user_operation_v07_mock = {
             "preVerificationGas": "0x1241e",
             "maxPriorityFeePerGas": "0x47868c00",
             "maxFeePerGas": "0x5dbefe60",
-            "factory": "0x4e1dcf7ad4e460cfd30791ccc4f9c8a4f820ec67",
+            "factory": "0x4e1DCf7AD4e460CfD30791CCC4F9c8a4f820ec67",
             "factoryData": "0x1688f0b9000000000000000000000000f9e93efbf37b588f9c79d509c48b87a26c3dfeb90000000000000000000000000000000000000000000000000000000000000060000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000844fff40e1277cc7ff46804f631ff0ed4cca5603a534066323b522073e113ee4d86d34298100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000060000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
             "signature": "0x000000000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000e04d29724cd758a76bde450eb4e2be71bf21398d55795de887cab857b85cd49a447d013104f681b1a225de0d97e4e3dfc8587861f847f539a05425591d3167c9f2000000000000000000000000000000000000000000000000000000000000002549960de5880e8c687434170f6476605b8fe4aeb9a28632c7995cf3ba831d976305000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000034226f726967696e223a22687474703a2f2f6c6f63616c686f73743a35313733222c2263726f73734f726967696e223a66616c7365000000000000000000000000",
         },
